gload.h: iterative loading
For loading process, I've resigned from loading everything line by line. Instead, I now have a key-value pair array of proc. names and function-pointer addressess which is then used by a for-loop to load every element of this array. By that, <<gload-loadfunc>> is no longer needed (for now it'll stay in the codebase, will probably remove it at some point). This solution works both for GLOAD_DLSYM and GLOAD_GLX so I expect it to work for the other backends.