gload.h: platform and version constants.
gload.h now defined a few new constants: GLOAD_LINUX, GLOAD_APPLE, GLOAD_WIN32 for detecting which platform is gload.h running on, GLOAD_PLATFORM as a string-literal platform representation and GLOAD_VERSION as a string-literal version representation of gload.h. With that addition there's now also a simple validation mechanism for the running system. With that tho, I should probably add another platform for BSD platforms, but that's for the future.